[#4] refactor: 공통 응답 dto 패키지 이동
