+++ module-core/src/main/java/com/soundie/global/common/dto/EnvelopeResponse.java
@@ -1,4 +1,4 @@
-package com.soundie.global.common;
+package com.soundie.global.common.dto;
 
 import lombok.*;
